refactor to remove adapter pattern
