--- lib/oli/grading/adapters/canvas.ex
@@ -1,241 +1,0 @@
-defmodule Oli.Grading.Adapters.Canvas do
-  @behaviour Oli.Grading.LtiGradeService
-
-  alias Oli.Delivery.Sections
-  alias Oli.Accounts
-
-  # Translators - These translate LTI 1.3 GS LineItems to LMS specific data structures.
-  # Some of the LTI and LMS parameters are commented out because they do not have a
-  # corresponding data value that matches the LTI standard.
-  defp translate_assignment_to_lineitem(assignment) do
-    %{
-      "id" => assignment["id"],
-      "label" => assignment["name"],
-      # "scoreMaximum" => number,
-      "tag" => assignment["name"],    # assignment name is the tag for this adapter
-      # "resourceId" => string,
-      # "resourceLinkId" => string,
-      # "submission" => {
-      #   "startDateTime" => string,
-      #   "endDateTime" => string,
-      # },
-    }
-  end
-
-  defp translate_lineitem_to_assignment(lineitem) do
-    %{
-      "id" => lineitem["id"],
-      "name" => lineitem["label"],
-      "submission_types" => [
-        "external_tool"
-      ],
-      "points_possible" => lineitem["scoreMaximum"],
-      "grading_type" => "points",
-      "published" => "true"
-    }
-  end
-
-  def product_family_code() do
-    # this adapter is associated with canvas platform product_family_code
-    "canvas"
-  end
-
-  def supported_methods() do
-    [
-      "get_lineitems",
-      "add_lineitem",
-      "add_lineitem_score",
-    ]
-  end
-
-  def basic_launch(lti_params) do
-    context_id = lti_params["context_id"]
-
-    case Sections.get_section_by(context_id: context_id) do
-      nil -> nil
-      section ->
-        # For canvas, we want to get the api_url using information from lti launch
-        %{protocol: protocol, host: host, port: port} = get_url_parts(lti_params["launch_presentation_return_url"])
-        api_url = "#{protocol}#{host}#{port}"
-        api_id = lti_params["custom_canvas_course_id"]
-
-        # update section record with api_url
-        Sections.update_section(section, %{api_url: api_url, api_id: api_id})
-    end
-  end
-
-  def get_lineitems(context_id, _query_params \\ %{}) do
-    %{api_url: api_url, api_token: api_token, api_id: api_id} = get_section(context_id)
-    url = "#{api_url}/api/v1/courses/#{api_id}/assignments"
-
-    fetch_all_pages(url, api_token)
-  end
-
-  def add_lineitem(context_id, lineitem) do
-    %{api_url: api_url, api_token: api_token, api_id: api_id} = get_section(context_id)
-    url = "#{api_url}/api/v1/courses/#{api_id}/assignments"
-
-    assignment = translate_lineitem_to_assignment(lineitem)
-    body = Jason.encode!(%{"assignment" => assignment})
-
-    authenticated_fetch(:post, api_token, url, body, ["content-type": "application/json; charset=utf-8"])
-  end
-
-  def get_lineitem(_context_id, _lineitem_id) do
-    throw "Method not supported"
-  end
-
-  def change_lineitem(_context_id, _lineitem_id, _lineitem) do
-    throw "Method not supported"
-  end
-
-  def remove_lineitem(_context_id, _lineitem_id) do
-    throw "Method not supported"
-  end
-
-  def get_lineitem_results(_context_id, _lineitem_id, _query_params \\ %{}) do
-    throw "Method not supported"
-  end
-
-  def add_lineitem_score(context_id, lineitem_id, score) do
-    assignment_id = get_lineitems(context_id)
-      |> Enum.find(fn lineitem -> lineitem["label"] ==  end)
-
-    # get lms_user_id from score userId
-    lms_user_id = Accounts.get_user_by(user_id: score["userId"])
-
-    %{api_url: api_url, api_token: api_token, api_id: api_id} = get_section(context_id)
-    url = "#{api_url}/api/v1/courses/#{api_id}/assignments/#{lineItemId}/submissions/#{submission.user_id}"
-    body = Jason.encode!(%{"submission" => %{ "posted_grade" => score["scoreGiven"]}})
-
-    authenticated_fetch(:put, token, url, body, ["content-type": "application/json; charset=utf-8"])
-  end
-
-  defp authenticated_fetch(method, api_token, url, body \\ "", headers \\ [], allow_token_refresh \\ true) do
-    # get an initial token if token is nil
-    token = case token do
-      nil -> refresh_token()
-      token -> token
-    end
-
-    full_headers = [
-      "Authorization": "Bearer #{token}",
-      "Accept": "application/json; charset=utf-8"]
-    ++ headers
-
-    result = case method do
-      :get ->
-        HTTPoison.get(url, full_headers)
-      :post ->
-        HTTPoison.post(url, body, full_headers)
-      :put ->
-        HTTPoison.put(url, body, full_headers)
-      :delete ->
-        HTTPoison.delete(url, full_headers)
-    end
-
-    case result do
-      {:ok, %HTTPoison.Response{status_code: 401} = response} when allow_token_refresh ->
-        # the response will contain WWW-Authenticate header to indicate a token refresh is required
-        case Enum.find(response.headers, fn {key, _value} -> key == "WWW-Authenticate" end) do
-          nil ->
-            {:ok, response}
-          _ ->
-            # get a new token and try again
-            token = refresh_token()
-            authenticated_fetch(method, token, url, body, headers, false)
-        end
-      other_result ->
-        other_result
-    end
-  end
-
-  defp refresh_token() do
-    # refresh token is not supported by this adapter yet. It assumes the token given is
-    # valid and does not expire
-    throw "API token is either missing or invalid. This adapter does not support token refresh."
-      <> "You must generate and configure a non-expiring token from canvas before using"
-  end
-
-  defp get_section(context_id) do
-    case Sections.get_section_by(context_id: context_id) do
-      nil ->
-        {:error, "Section does not exist for context_id #{context_id}"}
-      section ->
-        section
-    end
-  end
-
-  defp get_url_parts(url) do
-    captures = Regex.named_captures(~r/(?<protocol>https?)?[:\/\/]*(?<host>[^\/:]+)(?<port>:[^\/:]+)?/, url)
-    protocol = case captures do
-      %{"protocol" => protocol} ->
-        "#{protocol}://"
-      _ ->
-        ""
-    end
-    host = case captures do
-      %{"host" => host} ->
-        host
-      _ ->
-        ""
-    end
-    port = case captures do
-      %{"port" => port} ->
-        port
-      _ ->
-        ""
-    end
-
-    %{protocol: protocol, host: host, port: port}
-  end
-
-  # recursive function that fetches the page at the url given and checks the response
-  # for more pages. If another page is available, it recursively fetches that page until
-  # there are no more pages
-  defp fetch_all_pages(page_url, api_token, acc \\ []) do
-    case authenticated_fetch(:get, api_token, page_url) do
-      {:ok, %HTTPoison.Response{body: body, status_code: 200} = res} ->
-        assignments = body |> Jason.decode!
-        lineitems = Enum.map assignments, fn assignment -> translate_assignment_to_lineitem(assignment) end
-        acc = lineitems ++ acc
-
-        case next_page_url(res) do
-          nil ->
-            # no more pages to go, return acc
-            acc
-          next_url ->
-            fetch_all_pages(next_url, api_token, acc)
-        end
-      other_response ->
-        other_response
-    end
-  end
-
-  # parses the next page url in the response header, if it exists
-  defp next_page_url(res) do
-    case Enum.find(res.headers, fn {key, _} -> key == "link" end) do
-      nil -> nil
-      {_, link} ->
-        # parse next page link from Link header returned
-        rel_links = String.split(link, ",")
-        next_rel_link =
-          rel_links
-          |> Enum.map(fn rel_link ->
-            # e.g. <http://canvas.edu/api/v1/courses/1/custom_columns?page=2&per_page=10>; rel="next"
-            rel_link
-            |> String.split(";")
-            |> Enum.map(fn s -> String.trim(s) end)
-          end)
-          |> Enum.find(fn [_link, rel] -> rel == "rel=\"next\"" end)
-
-        case next_rel_link do
-          [link, _rel] ->
-            # strip off the leading < and trailing >
-            String.slice(link, 1, String.length(link) - 2)
-          _ -> nil
-        end
-    end
-  end
-
-end
