+++ lib/oli_web/controllers/lti_controller.ex
@@ -8,8 +8,7 @@ defmodule OliWeb.LtiController do
   alias Oli.Accounts.Institution
   alias Oli.Delivery.Sections
   alias Oli.Delivery.Sections.SectionRoles
-  alias Oli.Grading.Adapters
-  alias Oli.Grading.LtiGradeService
+  alias Oli.Grading.CanvasApi
 
   @doc """
   Handles an LTI basic launch
@@ -74,9 +73,9 @@ defmodule OliWeb.LtiController do
               conn
             end
 
-            # TODO: look up adapter type by section or institution, default to lti
-            adapter = Adapters.Canvas
-            adapter.basic_launch(conn.body_params)
+            # TODO: Remove when LTI 1.3 GS replaces canvas api for grade passback
+            # For now, we must capture some canvas specific information for grade passback
+            CanvasApi.handle_basic_launch(conn.body_params, user)
 
             # sign current user in and redirect to home page
             conn
