+++ lib/oli_web/router.ex
@@ -221,7 +221,13 @@ defmodule OliWeb.Router do
     get "/:context_id/page/:revision_slug/attempt/:attempt_guid", PageDeliveryController, :finalize_attempt
 
     get "/:context_id/grades/export", PageDeliveryController, :export_gradebook
-    get "/:context_id/grades/sync", PageDeliveryController, :sync_gradebook
+  end
+
+  scope "/api/v1/project", OliWeb do
+    pipe_through [:api, :protected, Oli.Plugs.SetCurrentUser]
+
+    post "/:context_id/grades/sync", PageDeliveryController, :sync_gradebook
+    put "/:context_id/grades/canvas_token", PageDeliveryController, :update_canvas_token
   end
 
   scope "/admin", OliWeb do
