+++ lib/oli_web/templates/page_delivery/index.html.eex
@@ -15,8 +15,60 @@
     <div class="flex-1"></div>
     <div>
       <%= link "Download Gradebook (.csv)", to: Routes.page_delivery_path(@conn, :export_gradebook, @context_id), class: "btn btn-link btn-sm" %>
-      <%= link "Sync Grades", to: Routes.page_delivery_path(@conn, :sync_gradebook, @context_id), class: "btn btn-warning btn-sm" %>
+
+      <div class="btn-group">
+        <button id="sync-grades" type="button" class="btn btn-warning btn-sm">Sync Grades</button>
+        <button id="configure-token" type="button" class="btn btn-outline-warning btn-sm" data-toggle="modal" data-target="#configure-token-modal">
+          <i class="fas fa-key"></i>
+        </button>
+      </div>
     </div>
+
+    <div class="modal fade" id="configure-token-modal" data-backdrop="static" tabindex="-1">
+      <div class="modal-dialog" role="document">
+        <div class="modal-content">
+          <div class="modal-header">
+            <h5 class="modal-title" id="staticBackdropLabel">Update Token</h5>
+            <button type="button" class="close" data-dismiss="modal" aria-label="Close">
+              <span aria-hidden="true">&times;</span>
+            </button>
+          </div>
+          <div class="modal-body">
+            <div class="input-group mb-3">
+              <div class="input-group-prepend">
+                <span class="input-group-text" id="basic-addon3">Token</span>
+              </div>
+              <input type="text" class="form-control" id="token">
+            </div>
+          </div>
+          <div class="modal-footer">
+            <button type="button" class="btn btn-secondary" data-dismiss="modal">Cancel</button>
+            <button id="save-token" type="button" class="btn btn-primary">Save</button>
+          </div>
+        </div>
+      </div>
+    </div>
+
+    <script>
+      $(function() {
+        $('#sync-grades').on('click', function() {
+          url = '<%= Routes.page_delivery_path(@conn, :sync_gradebook, @context_id) %>'
+          fetch(url, {
+            method: 'POST'
+          })
+        })
+        $('#save-token').on('click', function() {
+          token = $('token').text
+          url = '<%= Routes.page_delivery_path(@conn, :update_canvas_token, @context_id) %>'
+          fetch(url, {
+            method: 'PUT',
+            body: JSON.stringify({ token: token })
+          })
+
+          $('#configure-token-modal').modal('hide')
+        })
+      })
+    </script>
   </div>
   <% end %>
